requirements.txt
Flask==2.2.3
Werkzeug==2.2.3
Flask-Cors==3.0.10
pyjwt==2.6.0
marshmallow==3.19.0
sqlalchemy==2.0.7
colorama==0.4.6
psutil==5.9.5
celery==5.2.7
redis==4.5.4
requests==2.28.2
PyJWT==2.6.0
python-dotenv==1.0.0
gunicorn==20.1.0
email-validator==2.0.0
flask-caching==2.0.2
pytest==7.3.1
pytest-flask==1.2.0
pytest-cov==4.1.0 